# xxx-user-service/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /v1/permission/add:
        post:
            tags:
                - Permission
            operationId: Permission_Add
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.user.v1.PermSaveRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.PermReply'
    /v1/permission/list:
        get:
            tags:
                - Permission
            operationId: Permission_ListPage
            parameters:
                - name: pageNo
                  in: query
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  schema:
                    type: string
                - name: sort
                  in: query
                  schema:
                    type: string
                - name: direction
                  in: query
                  schema:
                    type: string
                - name: permName
                  in: query
                  schema:
                    type: string
                - name: url
                  in: query
                  schema:
                    type: string
                - name: action
                  in: query
                  schema:
                    type: string
                - name: permType
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.PermListReply'
    /v1/permission/update:
        post:
            tags:
                - Permission
            operationId: Permission_Update
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.user.v1.PermSaveRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.PermReply'
    /v1/role/add:
        post:
            tags:
                - Role
            operationId: Role_Add
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.user.v1.RoleSaveRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.RoleReply'
    /v1/role/assignPerm:
        post:
            tags:
                - Role
            operationId: Role_AssignPermission
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.user.v1.RoleAssignPermRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.RoleOk'
    /v1/role/code:
        get:
            tags:
                - Role
            operationId: Role_GetByCode
            parameters:
                - name: roleCode
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.RoleReply'
    /v1/role/delete:
        post:
            tags:
                - Role
            operationId: Role_DeleteById
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.user.v1.RoleDeleteRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.RoleOk'
    /v1/role/list:
        get:
            tags:
                - Role
            operationId: Role_ListPage
            parameters:
                - name: pageNo
                  in: query
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  schema:
                    type: string
                - name: sort
                  in: query
                  schema:
                    type: string
                - name: direction
                  in: query
                  schema:
                    type: string
                - name: roleCode
                  in: query
                  schema:
                    type: string
                - name: roleName
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.user.v1.RoleListReply'
...
components:
    schemas:
...
        api.user.v1.PermListReply:
            type: object
            properties:
                count:
                    type: string
                pageNo:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                permList:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.user.v1.PermReply'
        api.user.v1.PermReply:
            type: object
            properties:
                id:
                    type: string
                permName:
                    type: string
                url:
                    type: string
                action:
                    type: string
                permType:
                    type: integer
                    format: int32
                parentId:
                    type: string
        api.user.v1.PermSaveRequest:
            type: object
            properties:
                id:
                    type: string
                permName:
                    type: string
                url:
                    type: string
                action:
                    type: string
                permType:
                    type: integer
                    format: int32
                parentId:
                    type: string
        api.user.v1.RoleAssignPermRequest:
            type: object
            properties:
                RoleId:
                    type: string
                PermIdList:
                    type: array
                    items:
                        type: string
        api.user.v1.RoleDeleteRequest:
            type: object
            properties:
                id:
                    type: string
        api.user.v1.RoleListReply:
            type: object
            properties:
                count:
                    type: string
                pageNo:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                roleList:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.user.v1.RoleReply'
        api.user.v1.RoleOk:
            type: object
            properties:
                success:
                    type: boolean
        api.user.v1.RoleReply:
            type: object
            properties:
                id:
                    type: string
                roleCode:
                    type: string
                roleName:
                    type: string
        api.user.v1.RoleSaveRequest:
            type: object
            properties:
                id:
                    type: string
                roleCode:
                    type: string
                roleName:
                    type: string
...